+++ web/android/app/src/main/java/com/ghostzero/aistudycompanion/MainActivity.java
@@ -1,0 +1,5 @@
+package com.ghostzero.aistudycompanion;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}